+++ tests/Auth/Fixtures/AccessGateTestClass.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AccessGateTestClass
+{
+    public function foo()
+    {
+        return true;
+    }
+}
