+++ tests/Auth/Fixtures/AccessGateTestCustomResource.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AccessGateTestCustomResource
+{
+    public function foo($user)
+    {
+        return true;
+    }
+
+    public function bar($user)
+    {
+        return true;
+    }
+}
