+++ tests/Auth/Fixtures/AccessGateTestDummy.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AccessGateTestDummy implements AccessGateTestDummyInterface
+{
+    //
+}
