+++ tests/Auth/Fixtures/AccessGateTestDummyInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+interface AccessGateTestDummyInterface
+{
+    //
+}
