+++ tests/Auth/Fixtures/AccessGateTestPolicyWithBefore.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AccessGateTestPolicyWithBefore
+{
+    public function before($user, $ability)
+    {
+        return true;
+    }
+
+    public function update($user, AccessGateTestDummy $dummy)
+    {
+        return false;
+    }
+}
