+++ tests/Auth/Fixtures/AccessGateTestResource.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AccessGateTestResource
+{
+    public function view($user)
+    {
+        return true;
+    }
+
+    public function create($user)
+    {
+        return true;
+    }
+
+    public function update($user, AccessGateTestDummy $dummy)
+    {
+        return true;
+    }
+
+    public function delete($user, AccessGateTestDummy $dummy)
+    {
+        return true;
+    }
+}
