+++ tests/Auth/Fixtures/AccessGateTestSubDummy.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AccessGateTestSubDummy extends AccessGateTestDummy
+{
+    //
+}
