+++ tests/Auth/Fixtures/AuthTokenGuardTestUser.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class AuthTokenGuardTestUser
+{
+    public $id;
+
+    public function getAuthIdentifier()
+    {
+        return $this->id;
+    }
+}
