+++ tests/Auth/Fixtures/AuthorizesResourcesMiddleware.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+use Closure;
+
+class AuthorizesResourcesMiddleware
+{
+    public function handle($request, Closure $next, $method, $parameter)
+    {
+        return "caught can:{$method},{$parameter}";
+    }
+}
