+++ tests/Auth/Fixtures/EloquentProviderUserStub.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Tests\Auth\Fixtures;
+
+class EloquentProviderUserStub
+{
+    //
+}
